requirements.txt
django
gunicorn
djangorestframework
django-jazzmin